Added support for undefining features in a configuration. Added support for manual and automatic values in configuration.

+++ plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/conf/IConfigurationChanger.java
@@ -31,8 +31,8 @@ import de.ovgu.featureide.fm.core.Feature;
  */
 public interface IConfigurationChanger {
 
-	List<String> setFeature(Feature f, byte newValue);
+	List<String> setFeature(Feature f, int newValue);
 
-	void setNewValue(int index, byte value);
+	void setNewValue(int index, int value);
 
 }
