+++ plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/conf/nodes/Not2.java
@@ -9,7 +9,7 @@ public class Not2 extends Expression {
 	}
 
 	@Override
-	protected byte computeValue() {
+	protected int computeValue() {
 		switch (children[0].getValue()) {
 		case TRUE:
 			return FALSE;
